requirements.txt
requests==2.32.2
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.2
openai==1.30.4
python-dotenv==1.0.1
pyTelegramBotAPI==4.18.1
telebot==0.0.5
tensorflow==2.16.1
tensorflow_hub==0.16.1
tensorflow-text==2.16.1
scikit-learn==1.5.0
seaborn==0.13.2
matplotlib==1.16.0